setup roles

+++ backend/src/controllers/products.controller.js
@@ -3,23 +3,34 @@ import Product from '../models/producto.model.js';
 export const createProduct = async (req, res) => {
   try {
     const { name, sku, quantity, price } = req.body;
+
+    if (!req.user || !req.user.id) {
+      return res.status(401).json({ message: 'User ID is required' });
+    }
+
     const newProduct = new Product({
       name,
       sku,
       quantity,
       price,
-      vendedor: req.user.id,
+      user: req.user.id,
     });
     await newProduct.save();
-    res.json(newProduct);
+    res.status(201).json(newProduct);
   } catch (error) {
+    console.error('Error creating product:', error);
     return res.status(500).json({ message: error.message });
   }
 };
 
 export const getProducts = async (req, res) => {
   try {
-    const products = await Product.find({ vendedor: req.user.id }).populate(
+    if (req.user.role === 'administrador') {
+      const products = await Product.find().populate('user');
+      return res.json(products);
+    }
+
+    const products = await Product.find({ user: req.user.id }).populate(
       'vendedor',
     );
     res.json(products);
@@ -30,8 +41,15 @@ export const getProducts = async (req, res) => {
 
 export const getProduct = async (req, res) => {
   try {
-    const product = await Product.findById(req.params.id).populate('vendedor');
-    if (!product) return res.status(404).json({ message: 'Product not found' });
+    const product = await Product.findById(req.params.id).populate('user');
+
+    if (!product)
+      return res.status(404).json({ message: 'Producto no encontrado' });
+
+    if (product.user.toString() !== req.user.id) {
+      return res.status(403).json({ message: 'Acceso no autorizado' });
+    }
+
     return res.json(product);
   } catch (error) {
     return res.status(500).json({ message: error.message });
@@ -40,11 +58,16 @@ export const getProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   try {
-    const deleteProduct = await Product.findByIdAndDelete(req.params.id);
-    if (!deleteProduct)
-      return res.status(404).json({ message: 'Product not found' });
+    const product = await Product.findById(req.params.id);
 
-    return res.sendStatus(204);
+    if (!product) return res.status(404).json({ message: 'Product not found' });
+
+    if (product.user.toString() !== req.user.id) {
+      return res.status(403).json({ message: 'Unauthorized access' });
+    }
+
+    await Product.findByIdAndDelete(req.params.id);
+    res.sendStatus(204);
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -54,6 +77,15 @@ export const updateProduct = async (req, res) => {
   const { id } = req.params;
   const { name, sku, quantity, price } = req.body;
   try {
+    const product = await Product.findById(id);
+
+    if (!product)
+      return res.status(404).json({ essage: 'Producto no encontrado' });
+
+    if (product.user.toString() !== req.user.id) {
+      return res.status(403).json({ message: 'Acceso no autorizado' });
+    }
+
     const updatedProduct = await Product.findByIdAndUpdate(
       id,
       {
